+++ python/src/com/jetbrains/python/debugger/concurrency/tool/threading/PyThreadingLogManagerImpl.java
@@ -65,7 +65,7 @@ public class PyThreadingLogManagerImpl extends PyConcurrencyLogManager {
       }
     }
     PyConcurrencyEvent lastEvent = getEventAt(getSize() - 1);
-    int lastTime = lastEvent.getTime();
+    long lastTime = lastEvent.getTime();
     //set last time for stopping on a breakpoint
     for (ConcurrencyStat stat: result.values()) {
       if (stat.myFinishTime == 0) {
